Removed recommendation to export local functions
Originally, my thought was that exporting in-line would be best, but now I think it could be beneficial to organize all exports in one place to create a single view of the API for a given resource. Where the exports should live is context dependent and there isn't a one size fits all approach.